Adding a DB focused frontend, via a shared interface
